Color constructors. Context drawing. Window type.

+++ include/sengine/color.h
@@ -1,15 +1,18 @@
 #ifndef COLOR_H_
 #define COLOR_H_
 
+#include <stdint.h>
+
 namespace SEngine
 {
-    class Color {
-        public:
-            int r, g, b, a;
+    class Color
+    {
+    public:
+        uint8_t r, g, b, a;
 
-            Color();
-            Color(int r, int g, int b, int a);
-            Color(int r, int g, int b);
+        Color();
+        Color(uint8_t r, uint8_t g, uint8_t b, uint8_t a);
+        Color(uint8_t r, uint8_t g, uint8_t b);
     };
 } // namespace SEngine
